feat: 퍼센트 게이지 적용
