+++ lib/common/util.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:univ_note/portfolio/portfolio_main/common/caculater_d_day.dart';
+import '../portfolio/portfolio_main/common/set_default_start_or_end.dart';
 import '../user/register/model/user_information.dart';
 
 //유저 정보 가져오기.
@@ -15,7 +16,7 @@ Future<UserInformation> GetUserInformation() async {
       admissionDate: prefs.getString("admissionDate")??"",
       expectedGraduationDate: prefs.getString("expectedGraduationDate")??"",
       accessToken:prefs.getString("accessToken")??"",
-      Class_D_day:prefs.getString("Class_D_day")??SetClassDate()
+      classDday:prefs.getString("Class_D_day")??SetClassDate()
   );
   return user;
 }
