+++ lib/portfolio/portfolio_main/portfolio_screen.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:percent_indicator/linear_percent_indicator.dart';
 import 'package:univ_note/common/util.dart';
+import 'package:univ_note/portfolio/portfolio_main/common/caculater_percent.dart';
 import 'package:univ_note/portfolio/portfolio_main/component/d_day_button.dart';
 import 'package:univ_note/portfolio/portfolio_main/component/move_todoslist_button.dart';
 import 'package:univ_note/portfolio/portfolio_main/component/percent_view.dart';
@@ -12,6 +13,8 @@ import 'package:univ_note/setting/setting_main/component/profile_button.dart';
 import '../../common/basic.dart';
 import '../../user/register/model/user_information.dart';
 import 'common/caculater_d_day.dart';
+import 'common/find_start_or_end.dart';
+import 'common/set_start_dday.dart';
 
 class PortFolioScreen extends StatefulWidget {
   const PortFolioScreen({Key? key}) : super(key: key);
@@ -40,13 +43,13 @@ class _PortFolioScreenState extends State<PortFolioScreen> {
                         major: user.department,
                         end: user.expectedGraduationDate),
                     SizedBox(height: 10.h),
-                    D_DayButton(name: CaculaterClass(user.Class_D_day),clas: user.Class_D_day,grad: false,state: state),
+                    D_DayButton(name: CaculaterClass(user.classDday),clas: user.classDday,grad: false,state: state),
                     SizedBox(height: 10.h),
-                    PercentView(percent: 0.7),
+                    PercentView(percent: CaculaterPercent(SetStartAndPercent(user.classDday), user.classDday)),
                     SizedBox(height: 10.h),
                     D_DayButton(name: "졸업 🎓", clas: user.expectedGraduationDate,grad: true,state:state),
                     SizedBox(height: 10.h),
-                    PercentView(percent: 0.5),
+                    PercentView(percent: CaculaterPercent(user.admissionDate, user.expectedGraduationDate)),
                     SizedBox(height: 10.h),
                     Container(height: 2.h, color: Colors.black12),
                     SizedBox(height: 10.h),
